Add a join page and change GitHub to octocat logo
This commit adds a Presto Foundation page, with descriptions of membership
benefits and links to become a member.

It also includes the octocat logo, as there is no longer room in the header for
the word "GitHub". This commit includes changes to the menus for all existing
static pages, and the blog.

Signed-off-by: Brian Warner <brian@bdwarner.com>